Wave Intro Animation Fixed
I had forgotten to create a separate animation for Boss Wave intro, and overwrote my animation for the wave intro. This update fixes it.

+++ Assets/Scripts/PowerupManager.cs
@@ -34,7 +34,7 @@ public class PowerupManager : MonoBehaviour
         if (!canSpawnPowerup || playerPowerupLevel.value == playerMaxLevel)
             return;
         Debug.Log("Performing Random Check");
-        if (Random.Range(0,5*playerPowerupLevel.value) == 0)
+        if (Random.Range(0,5*(int)playerPowerupLevel.value) == 0)
         {
             Instantiate(powerupPrefab, position, powerupPrefab.transform.rotation, transform);
             canSpawnPowerup = false;
